@@ -1,16 +1,17 @@
-import {useState} from 'react'
-function Feedbackitem() {
-    const [rating, setRating] = useState(7);
-    const [text, setText] = useState('This is an example of feedback item');
-  
+
+
+const FeedbackItem = ({item}) => {
   return (
-    <div className="card">
-      <div className="num-display">{rating}</div>
-      <div className="text-display">{text}</div>
-     
-    </div>
+   <div className="card">
+    <div className="num-display">{item.rating}</div>
+    <div className="text-display">{item.text}</div>
+    <button>
+        
+    </button>
+   </div>
   )
 }
 
-export default Feedbackitem
+export default FeedbackItem
+
 
