@@ -1,47 +1,29 @@
 import PropTypes from 'prop-types'
+const Header = ({text, bgColor, textColor}) => {
 
-function Header({ text, bgColor, textColor }) {
-
-    const headerStyles = {
-        backgroundColor: bgColor, 
+    const headerStyle = {
+        backgroundColor  : bgColor,
         color: textColor
     }
   return (
-    <header style={headerStyles}>
-      <div className="container">
-          <h1>{text}</h1>
-      </div>
-    </header>
+   <header style ={headerStyle}>
+    <div className="container">
+        <h2>{text}</h2>
+    </div>
+   </header>
   )
 }
 
 Header.defaultProps = {
-    text : 'Feedback UI',
+    text: 'Feedback UI',
     bgColor : 'rgba(0,0,0,0.4)',
-    textColor : '#ff6a95'
+    textColor: '#ff6a95'
 }
 
 Header.propTypes = {
     text: PropTypes.string,
-    bgColor: PropTypes.string,
-    textColor: PropTypes.string
-
+    bgColor : PropTypes.string,
+    textColor : PropTypes.string
 }
 
 export default Header
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
